edgepipeline_register add ShouldDispatchToEdge method (#4693)
