+++ modules/pipeline/providers/edgepipeline/interface.go
@@ -16,14 +16,9 @@ package edgepipeline
 
 import (
 	"context"
-	"fmt"
-	"strings"
 
 	"github.com/erda-project/erda/apistructs"
-	"github.com/erda-project/erda/bundle"
 	"github.com/erda-project/erda/modules/pipeline/services/pipelinesvc"
-	"github.com/erda-project/erda/pkg/clusterdialer"
-	"github.com/erda-project/erda/pkg/discover"
 )
 
 type Interface interface {
@@ -35,47 +30,8 @@ func (p *provider) InjectLegacyFields(pipelineSvc *pipelinesvc.PipelineSvc) {
 	p.pipelineSvc = pipelineSvc
 }
 
-func (p *provider) ShouldDispatchToEdge(source, clusterName string) bool {
-	if clusterName == "" {
-		return false
-	}
-	if p.Cfg.ClusterName == clusterName {
-		return false
-	}
-	var findInWhitelist bool
-	for _, whiteListSource := range p.Cfg.AllowedSources {
-		if strings.HasPrefix(source, whiteListSource) {
-			findInWhitelist = true
-			break
-		}
-	}
-	if !findInWhitelist {
-		return false
-	}
-	isEdge, err := p.bdl.IsClusterDialerClientRegistered(clusterName, apistructs.ClusterDialerClientTypePipeline.String())
-	if !isEdge || err != nil {
-		return false
-	}
-	return true
-}
-
-func (p *provider) GetDialContextByClusterName(clusterName string) clusterdialer.DialContextFunc {
-	clusterKey := apistructs.ClusterDialerClientTypePipeline.MakeClientKey(clusterName)
-	return clusterdialer.DialContext(clusterKey)
-}
-
-func (p *provider) GetEdgeBundleByClusterName(clusterName string) (*bundle.Bundle, error) {
-	edgeDial := p.GetDialContextByClusterName(clusterName)
-	edgeDetail, err := p.bdl.GetClusterDialerClientData(apistructs.ClusterDialerClientTypePipeline.String(), clusterName)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get edge bundle for cluster %s, err: %v", clusterName, err)
-	}
-	pipelineAddr := edgeDetail.Get(apistructs.ClusterDialerDataKeyPipelineAddr)
-	return bundle.New(bundle.WithDialContext(edgeDial), bundle.WithCustom(discover.EnvPipeline, pipelineAddr)), nil
-}
-
 func (p *provider) CreatePipeline(ctx context.Context, req *apistructs.PipelineCreateRequestV2) (*apistructs.PipelineDTO, error) {
-	isEdge := p.ShouldDispatchToEdge(req.PipelineSource.String(), req.ClusterName)
+	isEdge := p.EdgePipelineRegister.ShouldDispatchToEdge(req.PipelineSource.String(), req.ClusterName)
 	if !isEdge {
 		pipeline, err := p.pipelineSvc.CreateV2(ctx, req)
 		if err != nil {
@@ -83,7 +39,7 @@ func (p *provider) CreatePipeline(ctx context.Context, req *apistructs.PipelineC
 		}
 		return p.pipelineSvc.ConvertPipeline(pipeline), nil
 	}
-	edgeBundle, err := p.GetEdgeBundleByClusterName(req.ClusterName)
+	edgeBundle, err := p.EdgePipelineRegister.GetEdgeBundleByClusterName(req.ClusterName)
 	if err != nil {
 		return nil, err
 	}
