+++ modules/pipeline/providers/edgepipeline_register/interface.go
@@ -19,11 +19,14 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"strings"
 	"time"
 
 	"github.com/rancher/remotedialer"
 
 	"github.com/erda-project/erda/apistructs"
+	"github.com/erda-project/erda/bundle"
+	"github.com/erda-project/erda/pkg/clusterdialer"
 	"github.com/erda-project/erda/pkg/discover"
 )
 
@@ -34,7 +37,50 @@ type Interface interface {
 	CheckAccessToken(token string) error
 	CheckAccessTokenFromHttpRequest(req *http.Request) error
 	IsEdge() bool
+
+	ShouldDispatchToEdge(source, clusterName string) bool
+	GetEdgeBundleByClusterName(clusterName string) (*bundle.Bundle, error)
 	//RegisterEdgeToDialer(ctx context.Context)
+}
+
+func (p *provider) ShouldDispatchToEdge(source, clusterName string) bool {
+	if clusterName == "" {
+		return false
+	}
+	if p.Cfg.ClusterName == clusterName {
+		return false
+	}
+	var findInWhitelist bool
+	for _, whiteListSource := range p.Cfg.AllowedSources {
+		if strings.HasPrefix(source, whiteListSource) {
+			findInWhitelist = true
+			break
+		}
+	}
+	if !findInWhitelist {
+		return false
+	}
+	isEdge, err := p.bdl.IsClusterDialerClientRegistered(clusterName, apistructs.ClusterDialerClientTypePipeline.String())
+	if !isEdge || err != nil {
+		return false
+	}
+
+	return true
+}
+
+func (p *provider) GetDialContextByClusterName(clusterName string) clusterdialer.DialContextFunc {
+	clusterKey := apistructs.ClusterDialerClientTypePipeline.MakeClientKey(clusterName)
+	return clusterdialer.DialContext(clusterKey)
+}
+
+func (p *provider) GetEdgeBundleByClusterName(clusterName string) (*bundle.Bundle, error) {
+	edgeDial := p.GetDialContextByClusterName(clusterName)
+	edgeDetail, err := p.bdl.GetClusterDialerClientData(apistructs.ClusterDialerClientTypePipeline.String(), clusterName)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get edge bundle for cluster %s, err: %v", clusterName, err)
+	}
+	pipelineAddr := edgeDetail.Get(apistructs.ClusterDialerDataKeyPipelineAddr)
+	return bundle.New(bundle.WithDialContext(edgeDial), bundle.WithCustom(discover.EnvPipeline, pipelineAddr)), nil
 }
 
 func (p *provider) GetAccessToken(req apistructs.OAuth2TokenGetRequest) (*apistructs.OAuth2Token, error) {
