+++ modules/pipeline/providers/edgepipeline_register/provider.go
@@ -29,6 +29,7 @@ import (
 type Config struct {
 	IsEdge                     bool          `env:"DICE_IS_EDGE" default:"false"`
 	ClusterName                string        `env:"DICE_CLUSTER_NAME"`
+	AllowedSources             []string      `env:"EDGE_ALLOWED_SOURCES"`
 	PipelineAddr               string        `env:"PIPELINE_ADDR"`
 	PipelineHost               string        `env:"PIPELINE_HOST"`
 	ClusterDialEndpoint        string        `file:"cluster_dialer_endpoint" desc:"cluster dialer endpoint"`
@@ -47,6 +48,7 @@ type provider struct {
 }
 
 func (p *provider) Init(ctx servicehub.Context) error {
+	p.bdl = bundle.New(bundle.WithClusterDialer())
 	if p.Cfg.IsEdge {
 		accessToken, err := os.ReadFile(p.Cfg.AccessTokenFile)
 		if err != nil {
